requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Zerodha API
kiteconnect==5.0.1

# Environment variables
python-dotenv==1.1.1
email-validator==2.1.1

# HTTP request
requests==2.32.5
httpx==0.27.0
google-auth==2.35.0

# Yahoo Finance for historical data
yfinance>=0.2.40
pandas>=2.0.0

# AWS Integration
boto3>=1.34.0
